pyproject.toml
[project]
name = "capstone-prompt-injection"
version = "0.1.0"
description = "Comparative evaluation of prompt injection defenses for enterprise AI agent deployments"
authors = [{ name = "Boglarka Petruska" }]
requires-python = ">=3.11,<3.13"

dependencies = [
    # Data and analysis
    "pandas>=2.0",
    "numpy>=2.0",
    "pyarrow>=15.0",
    "scikit-learn>=1.4",
    "scipy>=1.11",
    # Plotting
    "matplotlib>=3.8",
    "seaborn>=0.13",
    # HuggingFace
    "datasets>=3.0",
    "huggingface-hub>=0.24",
    "transformers>=4.40",
    "torch>=2.0",
    "safetensors>=0.4",
    "tokenizers>=0.20",
    "sentencepiece>=0.2",
    "protobuf>=4.0",
    # Notebook and utility
    "jupyter>=1.0",
    "ipykernel>=6.0",
    "python-dotenv>=1.0",
    "tqdm>=4.65",
    "tabulate>=0.9",
]

[project.optional-dependencies]
# API clients, installed when Defense B work begins
api = [
    "anthropic>=0.40",
    "openai>=1.50",
    "groq>=0.10",
]

[tool.uv]
package = false
